todo_api/api/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import generics
from rest_framework.permissions import IsAuthenticated

from todo_api.api import facade
from todo_api.api.filters import TaskFilter
from todo_api.api.serializers import TaskSerializer


class TaskListCreateView(generics.ListCreateAPIView):
    filter_backends = [DjangoFilterBackend]
    filterset_class = TaskFilter
    permission_classes = [IsAuthenticated]
    serializer_class = TaskSerializer

    def get_queryset(self):
        return facade.tasks(self.request.user)
