requirements.txt
asgiref==3.4.1
attrs==21.2.0
dj-database-url==0.5.0
Django==3.2.8
django-extensions==3.1.3
django-filter==21.1
django-health-check==3.16.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
iniconfig==1.1.1
gunicorn==20.1.0
packaging==21.0
pluggy==1.0.0
psutil==5.8.0
py==1.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-decouple==3.5
psycopg2==2.9.1
pytz
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1